requirements.txt
langchain
huggingface_hub
transformers
torch
torchvision
torchaudio
chainlit
pypdf
gpt4all
google-search-results
pinecone-client
langchain-openai
tiktoken
wikipedia
faiss-cpu
numexpr
